Title: 숫자 사이의 배수, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240312/숫자 사이의 배수/multiplication-between-numbers.c
@@ -13,7 +13,7 @@ int main() {
         }
     }
 
-    ave = sum / cnt;
+    ave = (float)sum / cnt;
 
     printf("%d %.1f",sum, ave);
     return 0;
